modifié :         include/calculations/addition.h 	modifié :         include/calculations/multiplication.h 	modifié :         include/calculations/substraction.h 	modifié :         include/gradient_descent.h 	modifié :         srcc/calculations/addition.c 	modifié :         srcc/calculations/multiplication.c 	modifié :         srcc/calculations/substraction.c 	modifié :         srcc/gradient_descent.c 	modifié :         srcc/main.c

+++ include/calculations/addition.h
@@ -1,6 +1,6 @@
 #ifndef __ADDITION_H__
 #define __ADDITION_H__
     
-    int* add(const int *mat1, const int *mat2, const int size);
+    float* add(const float *mat1, const float *mat2, const int size);
 
 #endif //__ADDITION_H__
