+++ include/calculations/substraction.h
@@ -1,6 +1,6 @@
 #ifndef __SUBSTRACTION_H__
 #define __SUBSTRACTION_H__
     
-    int* sub(const int *mat1, const int *mat2, const int size);
+    float* sub(const float *mat1, const float *mat2, const int size);
 
 #endif //__SUBSTRACTION_H__
