+++ srcc/calculations/multiplication.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int* mul(const int *mat1, const int *mat2, const int *column_result, const size_t size1, const size_t size2) {
+float* mul(const float *mat1, const float *mat2, const int *column_result, const int size1, const int size2) {
 
     int number_row_mat2 = size2 / (*column_result);
     //fprintf(stderr, "\nnumber_row_mat2 = size2 / (*column_result) : %d = %zu / %d\n", number_row_mat2, size2, (*column_result));
@@ -11,7 +11,7 @@ int* mul(const int *mat1, const int *mat2, const int *column_result, const size_
 
     // the number of rows of mat1 is equal to the number of 
     // rows of the matrix as the result
-    int *mat_result = (int*)malloc(((*column_result) * number_row_mat1) * sizeof(int));
+    float *mat_result = (float*)malloc(((*column_result) * number_row_mat1) * sizeof(float));
     //printf("\nle resultat doit contenir %d nombres", sizeof(*mat_result));
 
     for (int row_mat1 = 0; row_mat1 < number_row_mat1; row_mat1++) {
@@ -20,7 +20,7 @@ int* mul(const int *mat1, const int *mat2, const int *column_result, const size_
         // equal to the number of columns of the matrix as the result :
         for (int col_mat2 = 0; col_mat2 < *column_result; col_mat2++) {
 
-            int result = 0;
+            float result = 0;
 
             for (int row_mat2 = 0; row_mat2 < number_row_mat2; row_mat2++) {
