+++ srcc/calculations/substraction.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int* sub(const int *mat1, const int *mat2, const int size) {
+float* sub(const float *mat1, const float *mat2, const int size) {
 
-    int *vec_result = (int*)malloc(size * sizeof(int));
+    float *vec_result = (float*)malloc(size * sizeof(float));
 
     for(int vector_iterator = 0; vector_iterator < size; vector_iterator++) {
